Move StateMachine's abstraction from the Core to Sylius Contracts package
