+++ src/Sylius/Bundle/CoreBundle/EventListener/Workflow/Order/CancelShipmentListener.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\EventListener\Workflow\Order;
 
-use Sylius\Bundle\CoreBundle\StateMachine\StateMachineInterface;
+use Sylius\Component\Contracts\StateMachine\StateMachineInterface;
 use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Shipping\ShipmentTransitions;
 use Symfony\Component\Workflow\Event\CompletedEvent;
