+++ src/Sylius/Bundle/CoreBundle/Twig/StateMachineExtension.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\Twig;
 
-use Sylius\Bundle\CoreBundle\StateMachine\StateMachineInterface;
+use Sylius\Component\Contracts\StateMachine\StateMachineInterface;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
