+++ src/Sylius/Bundle/CoreBundle/spec/EventListener/Workflow/Order/CancelPaymentListenerSpec.php
@@ -15,7 +15,7 @@ namespace spec\Sylius\Bundle\CoreBundle\EventListener\Workflow\Order;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use PhpSpec\ObjectBehavior;
-use Sylius\Bundle\CoreBundle\StateMachine\StateMachineInterface;
+use Sylius\Component\Contracts\StateMachine\StateMachineInterface;
 use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Core\Model\PaymentInterface;
 use Sylius\Component\Payment\PaymentTransitions;
