+++ src/Sylius/Component/Contracts/StateMachine/TransitionInterface.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\StateMachine;
+namespace Sylius\Component\Contracts\StateMachine;
 
 interface TransitionInterface
 {
